Introduce SqliteExtendedErrorCode and centralize SQLite error handling

+++ CiccioSoft.Data.Sqlite/SqliteCommand.cs
@@ -185,7 +185,7 @@ public class SqliteCommand : DbCommand
             }
             catch (SqliteInteropException ex) when (_timeoutTriggered && ex.BaseErrorCode == SqliteResult.Interrupt)
             {
-                throw new SqliteException(Properties.Resources.CommandTimedOut(_command.CommandTimeout), (int)SqliteResult.Interrupt, ex.ExtendedErrorCode, ex);
+                throw new SqliteException(Properties.Resources.CommandTimedOut(_command.CommandTimeout), (int)SqliteResult.Interrupt, (int)ex.ExtendedErrorCode, ex);
             }
             catch (SqliteInteropException ex) when ((operationCanceled || operationCancellationToken.IsCancellationRequested) && ex.BaseErrorCode == SqliteResult.Interrupt)
             {
@@ -197,7 +197,7 @@ public class SqliteCommand : DbCommand
             }
             catch (SqliteInteropException ex)
             {
-                throw new SqliteException(ex.Message, (int)ex.BaseErrorCode, ex.ExtendedErrorCode, ex);
+                throw new SqliteException(ex.Message, (int)ex.BaseErrorCode, (int)ex.ExtendedErrorCode, ex);
             }
             finally
             {
